session_catalog: carve detail_search.rs (pure move)
Session detail responses, context-snapshot selectors, and the session
log search (terms, candidates, snippets, filters) — with tests,
verbatim. Purity-checked; tests conserved at 104.
